+++ include/CMakeLists.txt
@@ -3,9 +3,13 @@ add_subdirectory(carla_osi)
 set(CARLA_OSI_LIB_INCLUDES ${CARLA_OSI_LIB_INCLUDES}
 ${CMAKE_CURRENT_LIST_DIR}/CARLA_OSI_Interface.h
 ${CMAKE_CURRENT_LIST_DIR}/CARLA_TrafficUpdate.h
+${CMAKE_CURRENT_LIST_DIR}/CARLA_GroundTruth.h
+${CMAKE_CURRENT_LIST_DIR}/CARLA_SensorView.h
+${CMAKE_CURRENT_LIST_DIR}/Logger.h
 ${CMAKE_CURRENT_LIST_DIR}/Utility.h
 ${CMAKE_CURRENT_LIST_DIR}/ParameterDefinitions.h
 ${CMAKE_CURRENT_LIST_DIR}/Semaphore.h
 ${CMAKE_CURRENT_LIST_DIR}/CARLA_Interface.h
+${CMAKE_CURRENT_LIST_DIR}/CARLA_Module.h
 PARENT_SCOPE
 )
