+++ src/CMakeLists.txt
@@ -6,7 +6,10 @@ set(CARLA_OSI_LIB_SOURCE ${CARLA_OSI_LIB_SOURCE}
 ${CMAKE_CURRENT_LIST_DIR}/CARLA_Interface.cpp
 ${CMAKE_CURRENT_LIST_DIR}/CARLA_OSI_Interface.cpp
 ${CMAKE_CURRENT_LIST_DIR}/CARLA_TrafficUpdate.cpp
-${CMAKE_CURRENT_LIST_DIR}/Utility.cpp 
+${CMAKE_CURRENT_LIST_DIR}/CARLA_GroundTruth.cpp
+${CMAKE_CURRENT_LIST_DIR}/CARLA_SensorView.cpp
+${CMAKE_CURRENT_LIST_DIR}/Logger.cpp
+${CMAKE_CURRENT_LIST_DIR}/Utility.cpp
 
 PARENT_SCOPE
 )
